Fix thermostat mode to NOT trigger climate start/stop
Problem: The auto() and off() thermostat mode commands were triggering
StartClimate() and StopClimate() automatically, which goes against the design
decision to decouple configuration from action.

Issue: Users need to be able to configure ALL settings (temperature, defrost,
steering wheel, seats, etc.) before starting climate. If changing thermostat
mode immediately starts climate, users can't configure other settings first.

Solution: Thermostat mode commands now ONLY update the mode attribute without
triggering any climate actions:
- auto() - sets mode to 'auto' (configuration only)
- off() - sets mode to 'off' (configuration only)
- setThermostatMode() - updates mode attribute only

Climate control workflow:
1. User configures temperature via thermostat tile
2. User configures defrost switches (child devices)
3. User configures heated steering wheel (child device)
4. User configures heated/cooled seats (child devices)
5. User explicitly presses StartClimate button to start
6. User presses StopClimate button to stop

This prevents accidental API calls and gives users full control over when
climate commands are sent to the vehicle.

+++ OpenEVSEDriver.groovy
@@ -188,48 +188,48 @@ def parseStatusResponse(data) {
     
     // Charging current and voltage (amp is in mA, voltage and pilot are already in V and A)
     if (data.amp != null) {
-        def current = (data.amp / 1000).round(2)  // Convert mA to A and round to 2 decimal places
+        def current = (data.amp.toFloat() / 1000).round(2)  // Convert mA to A and round to 2 decimal places
         sendEvent(name: "chargingCurrent", value: current, unit: "A")
         sendEvent(name: "amperage", value: current)  // CurrentMeasurement capability
     }
     
     if (data.pilot != null) {
-        def maxCurrent = data.pilot.round(2)  // Already in A, just round to 2 decimal places
+        def maxCurrent = data.pilot.toFloat().round(2)  // Already in A, just round to 2 decimal places
         sendEvent(name: "maxCurrent", value: maxCurrent, unit: "A")
     }
     
     if (data.voltage != null) {
-        def voltage = data.voltage.round(2)  // Already in V, just round to 2 decimal places
+        def voltage = data.voltage.toFloat().round(2)  // Already in V, just round to 2 decimal places
         sendEvent(name: "voltage", value: voltage, unit: "V")
     }
     
     // Power calculation (amp is in mA, voltage is in V)
     if (data.amp != null && data.voltage != null) {
-        def power = ((data.amp / 1000) * data.voltage / 1000).round(2)  // Convert mA to A, multiply by V, then to kW and round to 2 decimal places
+        def power = ((data.amp.toFloat() / 1000) * data.voltage.toFloat() / 1000).round(2)  // Convert mA to A, multiply by V, then to kW and round to 2 decimal places
         sendEvent(name: "power", value: power, unit: "kW")
     }
     
     // Energy
     if (data.session_wh != null) {
-        def sessionEnergy = (data.session_wh / 1000).round(2)  // Convert Wh to kWh and round to 2 decimal places
+        def sessionEnergy = (data.session_wh.toFloat() / 1000).round(2)  // Convert Wh to kWh and round to 2 decimal places
         sendEvent(name: "sessionEnergy", value: sessionEnergy, unit: "kWh")
     }
     
     if (data.watthour != null) {
-        def totalEnergy = (data.watthour / 1000).round(2)  // Convert Wh to kWh and round to 2 decimal places
+        def totalEnergy = (data.watthour.toFloat() / 1000).round(2)  // Convert Wh to kWh and round to 2 decimal places
         sendEvent(name: "totalEnergy", value: totalEnergy, unit: "kWh")
         sendEvent(name: "energy", value: totalEnergy)  // EnergyMeter capability
     }
     
     // Session time
     if (data.elapsed != null) {
-        def sessionMinutes = (data.elapsed / 60).round(2)  // Convert seconds to minutes and round to 2 decimal places
+        def sessionMinutes = (data.elapsed.toFloat() / 60).round(2)  // Convert seconds to minutes and round to 2 decimal places
         sendEvent(name: "sessionTime", value: sessionMinutes, unit: "min")
     }
     
     // Temperature
     if (data.temp != null) {
-        def tempCelsius = (data.temp / 10).round(2)  // Convert deci-degrees to Celsius and round to 2 decimal places
+        def tempCelsius = (data.temp.toFloat() / 10).round(2)  // Convert deci-degrees to Celsius and round to 2 decimal places
         sendEvent(name: "temperature", value: tempCelsius, unit: "°C")
     }
